Create aws infra subcommand for aws-specific infra
